added ownership

+++ apps/gateway/src/services/users/users.service.ts
@@ -1,4 +1,10 @@
-import { Body, Inject, Injectable, OnModuleInit } from '@nestjs/common';
+import {
+  Body,
+  Inject,
+  Injectable,
+  OnModuleInit,
+  UnauthorizedException,
+} from '@nestjs/common';
 import {
   AuthenticateRequest,
   CreateUserRequest,
@@ -19,19 +25,34 @@ export class UsersService implements OnModuleInit {
       this.client.getService<UsersServiceClient>(USERS_SERVICE_NAME);
   }
 
+  async checkOwnership(userId: string, jwtUserId: string) {
+    const user = await this.usersService.getUser({ id: userId }).toPromise();
+
+    if (user?.id !== jwtUserId) {
+      throw new UnauthorizedException(
+        'You dont have permission to access this resource',
+      );
+    }
+
+    return;
+  }
+
   createUser(@Body() createUserDto: CreateUserRequest) {
     return this.usersService.createUser(createUserDto);
   }
 
-  getUser(id: string) {
+  async getUser(id: string, jwtUserId: string) {
+    await this.checkOwnership(id, jwtUserId);
     return this.usersService.getUser({ id });
   }
 
-  updateUser(updateUserDto: UpdateUserRequest) {
+  async updateUser(updateUserDto: UpdateUserRequest, jwtUserId: string) {
+    await this.checkOwnership(updateUserDto.id, jwtUserId);
     return this.usersService.updateUser(updateUserDto);
   }
 
-  deleteUser(id: string) {
+  async deleteUser(id: string, jwtUserId: string) {
+    await this.checkOwnership(id, jwtUserId);
     return this.usersService.deleteUser({ id });
   }
